refactor: moved product tests to a domain folder
