+++ .prettierrc
@@ -1,3 +1,4 @@
 {
-  "singleQuote": false
+  "singleQuote": false,
+  "trailingComma": "none"
 }
